cms/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
Django==4.2.11
django-crispy-forms==2.1
django-tinymce==4.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
pillow==10.2.0
sqlparse==0.4.4
typing-extensions==4.10.0
